The problem of distinguishing prime numbers from composite numbers and of resolving the latter into their prime factors is known to be one of the most important and useful in arithmetic

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -14,7 +14,7 @@ int main(void)
 	int count;
 	int n;
 
-	int number = 612852475143;
+	long number = 612852475143;
 
 	for (count = 2; count <= number; count++)
 	{
